auth-service/requirements.txt
fastapi==0.115.0
pydantic==2.9.2
PyJWT==2.9.0
uvicorn==0.31.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-exporter-otlp==1.27.0
python-ulid==3.0.0